requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pymupdf==1.24.7
python-dotenv==1.0.1
python-docx==1.1.2
openpyxl==3.1.5
beautifulsoup4==4.12.3
